[QD-8] Small refactoring and fixes
