+++ server/src/common/logger/index.js
@@ -2,11 +2,11 @@ const winston = require('winston');
 const loggerConfig = require('../../config').logger;
 
 const logger = new (winston.Logger)({
-    exitOnError: false,
-    transports: [
-        new (winston.transports.Console)(loggerConfig.transports.console),
-        new (winston.transports.File)(loggerConfig.transports.file),
-    ],
+  exitOnError: false,
+  transports: [
+    new (winston.transports.Console)(loggerConfig.transports.console),
+    new (winston.transports.File)(loggerConfig.transports.file),
+  ],
 });
 
 module.exports = logger;
