+++ server/src/constants/index.js
@@ -1,5 +1,3 @@
 const errors = require('./errors');
 
-module.exports = {
-  errors,
-};
+module.exports = { errors };
