+++ server/src/models/index.js
@@ -1,0 +1,5 @@
+const question = require('./question');
+const quiz = require('./quiz');
+const result = require('./result');
+
+module.exports = { question, quiz, result };
